requirements.txt
asgiref==3.5.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
charset-normalizer==2.1.1
click==8.1.3
cloudinary==1.30.0
constantly==15.1.0
cryptography==38.0.1
daphne==4.0.0
Django==4.1.2
django-anymail==8.6
django-autoslug==1.9.8
django-cloudinary-storage==0.3.0
django-debug-toolbar==3.7.0
django-jazzmin==2.5.0
django-ninja==0.19.1
django-sms==0.5.0
gunicorn==20.1.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
Pillow==9.2.0
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
PyJWT==2.6.0
pyOpenSSL==22.1.0
python-decouple==3.6
pytz==2022.5
r-django-jet==2.0.9
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
twilio==7.15.0
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
whitenoise==6.2.0
zope.interface==5.5.0
